Correct linking for each target.
Now that most of the code is in a shared library, there's no need to
attempt to link with all its dependencies in each of the executables.
Additionally, it's incorrect to place -llib in linker flags because that
goes before object files, and not after, which could cause problems with
the -Wl,--as-needed flag.
Setting -Wl,--version-script for all shared libraries is wrong, because
the Firefox plugin does not include all of that code any more.
There were also some unnecessary libraries added to the Firefox plugin
link command.