Fixes in geom_Matrix test.
Add matrixEqual() helper to compare contents of matrices, the regular
comparison is for object identity and equality test fails.
Matrix.concat() works inplace and returns NULL, do not use it as an
expression in tests.
Fix expected values for concat and rotate tests.
The invert test still fails even if now the result is ok, because
inequality of -0.00 and +0.00