Remove GraphicsContainer, add TokenContainer
GraphicsContainer is gone, Shape and Sprite have now a _NR(Graphics)
member. The token list was removed from Graphics and is now
in TokenContainer. Also in TokenContainer are all functions that
operate on a list of tokens, such as getBounds(), invalidate()
hit() test etc. Shape and Sprite are subclasses of TokenContainer.
The Graphics object gets a pointer to a TokenContainer where it
draws on.