Purge Frame and handle everything in the 'dynamic' display list
After this commit, the Frame class does only contain a list
of DisplayListTags, the blueprints and one method to apply
those blueprints to a given DisplayObjectContainer.
MovieClip::advanceFrame() will now just execute each
frame between the current and the next to bring itself
(which is a DisplayObjectContainer) to the current state.
We can remove frame.h/cpp, because no more functionality is needed.
We can remove all second paths in flashdisplay that were of the form
MovieClip::Render() {
	Sprite::Render();
	for each obj in current frame do Render
}
becomes just Sprite::Render(), because all objects in the frame
are now childs of the DisplayObjectContainer.