Pygil: Fix parsing of static properties, print warnings
Pygil will now warn if lightsparks implements a getter/setter/method
which is not in the spec. This matches strictly to classes,
i.e. properties must be provides by exact the same class
and not a subclass. Example: Currently you get a warning
"Warning: Implemented non-spec getter flash.display.Stage.loaderInfo"
because loaderInfo must be provided by DisplayObject.

See http://fpaste.org/xzf5/ for the complete output of the script.