Make parser/vm interaction safer (and faster)
Removes the state.max_FP as it is the same as framesLoaded.

Introduces a Spinlock for access to framesLoaded and wrap
all access into getFramesLoaded() and setFramesLoaded().

Makes frames a std::list, so that no additional locking for
access is required. Removed framesMutex.

Rename totalFrames to totalFrames_unreliable and remove
all usage of it, but in AS getter functions, so its just informational.

Make frameScripts a map, so we do not have to resize it
when another frame gets added. This also saves some space,
as it is usually sparse.

One missing touch point of parser/vm is setOnStage,
which will be handled in a later commit.