HACK: Make the display work on GLES2 by not chunking textures.
Instead of many 128x128 textures use a 1024x1024 one for now.
This makes GLES2 display the content without artifacts caused by the
lack of fallback code for pixel unpack alignment and mapping of the
pixel buffers with PBOs in OpenGL.
This needs fixing, but for now allows testing the GLES2 codepaths.