Add basic support for MouseMove, MouseOver, MouseOut and DoubleClick mouse events.
For now the hitTest/hitTestImpl mechanism used to decide which object
should receive a Click event is reused for the events mentioned above.
The InteractiveObject.doubleClickEnabled property is not supported yet.
Another effect of this commit is to disable passing on mouse click events
originating from the right mouse button.

The correctness of this implementation w.r.t. the specifications is not yet fully
tested and thus the implementation is probably not fully correct.

Further improvements will follow.