Add support for InteractiveObject.doubleClickEnabled and correctly implement InteractiveObject.mouseEnabled.
The DisplayObject::hitTest and DisplayObject::hitTestImpl methods
now accepts an additional "type" parameter.
This parameter defines what type of hit must be tested for.
For now, the two possibilities are GENERIC_HIT and DOUBLE_CLICK.
GENERIC_HIT is used for any hit-test with an unspecified type.
DOUBLE_CLICK is used for hit-tests for double-mouse-clicks.

Any subclass of DisplayObjectContainer should use DisplayObjectContainer's
hitTestImpl. This method will also check if the container should accept mouse
and/or double-click events. This method will return it's given "last" argument
if-and-only-if none of its children are hit and only the DisplayObjectContainer
itself remains as a potential candidate.
The subclass can then decide for itself if it is really hit or not.