Re-registration of an event listener should be ignored
Changes to the old behaviour: 1) When checking if the listener has
already been added, search through all priorities, don't just compare
to the successor of the insertion point. 2) Use listener's ==
operator (or rather find()) instead of comparing IFunction pointers
(an AS function can be represented by multiple IFunction instances).

The practical need for this is the odd behaviour in flowplayer where
every MOUSE_MOVE event tries to re-add the listener. Without this
patch each event effectively doubles the number of listeners, which
causes bad performance problems.