PulsePlugin: Use multiple writes to fill buffer
This is necessary as pa_stream_begin_write does not return the whole
available buffer at once, but in sequential calls. It makes
the code more efficient as less callbacks from pulse are necessary
to fill the buffer.
And it fixes a bug on my platform where I have no sound at all.