Implement ABCVm::constructGenericType()
It allows to instantiate a class from a template T and template
arguments S.

In the implemented language:
Given a Template<T> and some Class<S> on the stack, it returns a
TemplatedClass<T> (which is a
subclass of Class<T>) with 'types' member set to Class<S>.
This class can then be constructed by the "construct" op.

This allows to handle the "Vector" template.