Change semantics of Class_base::getClass()
There were two usages of getClass() in the code,
one which expected getClass() to do incRef() on the class
before returning it and one that expected getClass() to not
incRef it before.

Expected incRef:
 c->super = Class<ASObject>::getClass()
 builtin->setVariableByQName("TextField","",Class<ASObject>::getClass())

Expected no incRef:
 if(c->prototype == Class<ASObject>::getClass())
 if(c->prototype->isSubClass(Class<ASObject>::getClass()))

Previous commits have introduced Class_base::getRef() which
returns a Ref<Class_base> to make obvious that the caller gets
a fresh reference. Previous commits also have
converted the usecases 'super' and 'builtin->setVariableByQName'
to use getRef().
This commit now changes the semantics of getClass() to return
the Class_base object _without_ incRef'ing it in accordance
to the two other usecases ('prototype == Class<ASObject>::getClass()'
and 'isSubClass').