getLex should skip XML and XMLList objects which do no have the requested property
When the new XML_STRICT flag is set in getVariableByMultiname call,
the call will ignore non-existing properties in XML objects. Without
the flag, trying to get a non-existing property returns an empty
XMLList.

This fixes a regression in the flowplayer demo by restoring (a
rewritten version of) the previously reverted commit. The commit was
reverted because it exposed a bug in PlaceObejct tag handling, but
that has now been fixed by 48a5dc373ad777343343ce79fbb87fc6389c9b34.