Add copy of NPAPI headers to source
Header files required to build the browser plugin taked from the firefox-dev
package corresponding to Firefox 7 as shipped in Ubuntu.There is only a
slight difference in a couple of function signatures compared to the
headers as shipped by the npapi-sdk project (const char * vs char *).
Changing to those can happen in the future.