Implement the ARG_UNPACK macro
Usage of ARG_UNPACK:
You have to use it within a ASFUNCTIONBODY() { }, because it uses
the implicit arguments 'args' and 'argslen'.
Standard usage is:
int32_t i;
bool b;
_NR<DisplayObject> o;
ARG_UNPACK (i) (b) (o);
which coerces the given arguments implicitly (according to ecma) to
the types Integer, Boolean, DisplayObject and then puts them
into the given variables.
ATTENTION: The object 'o' is the same as the argument passed into
the function, so changing it will be visible for the caller.

A exception will be thrown if a type cannot be coerces implicitly or
when not enough arguments are provided by the caller. A
LOG_NOT_IMPLEMENTED
will be emitted if more than the unpacked arguments are provided by
the caller.

To specify default values, use the (var, defvalue) operator like
ARG_UNPACK (i) (b,true) (o,_NR(new Undefined));
this will work as above if all arguments are supplied. When the
second argument is not supplied, no error is thrown and
b is set to 'true'. If the third argument is not supplied, not error
is thrown and o is set to _NR(new Undefined).