Resolve slot type on access, not on creation
This fixes problems with definitions like
class A
{
  var b:B;
}
class B
{
  var a:A;
}

In the script init, first class A is created. During creating of
its slot 'b', it encounters the T_DEFINABLE class B. It can only be
defined by calling the script init it belongs to, but that one is
already running.
We now defer definition of T_DEFINABLE class B until one actually
write something other than 'null' to 'b'.