Definable::define: Look in the classes map
This is necessary when in code like
Class A {
  static var a:A = initMe();
  function initMe():A { return new A; }
}
which translates to
  newClass A
  setProp A
During the opcode newClass, we call the static class initalizer
which calls initMe() to initialize the variable 'a'.
We need to resolve its type 'A' here, but the global variable
'A' still points to a Definable. (It is changed to point to the new
class in the opcode 'setProp A')