Use Type for the type of slot traits, fix coercing in setVar
For late resolving of types, we add a multiname* typemname to variable.
On the first setVar(), typemname is resolved and cached for later
use. setVar then uses type->coerce() to convert the value.
The previously used generate() is not correct, as it performs
_explicit_ type conversion.

This commit also removes the use of Definable for the value
of Slot and Const traits. As those traits can only be accessed
after their class is defined (i.e. their class's static init has
run), their values will already be defined.

This simplifes the use of Definable, as it now is only a placeholder
for Classes. Therefore resolving a Definable amounts to a search
in the SystemState::classes map.