Implement Namespace.valueOf/Namespace.prototype.valueOf
Namespace.valueOf differs from Namespace.prototype.valueOf because
the second one has to check if the object it is invoked on is actually
a Namespace. (See also tamarin/core/Namespace.as)
The Namespace.prototype.valueOf is a relict from ECMA (therfore its in
the "" namespace). AS3 prefers the Namespace.valueOf method (which is in
the AS3 namespace).

This also makes the Namespace.prototype.valueOf != Namespace.valueOf
check succed that some of the tests use.