Convert 'cur_level' and friends to 'inClass'
We used a quite big machinery of getLevel()/resetLevel() etc.
for just the four opcodes {set,get,call,construct}Super.

Now, we have IFunction::inClass, which is a pointer to the class
which this is a method in.
In SyntheticFunction::callImpl, IFunction::inClass is copied into
call_context::inClass.
The opcodes than use call_context::inClass->super to look for the
necessary traits.

This removes and simplifies a lot of code.