[ExternalInterface] Make sure conversions from ASObject to ExtVariant and the other way around are done on the VM thread.
To support this behavior I have implemented a new type of WaitableEvent: ExternalCallEvent.
This type of event handles a call from Javascript to a callback function registered through ExternalInterface.
It takes arguments to the callback in the form of ExtVariants and return the result as an ExtVariant.
This means that the ExtVariants are actually converted to ASObjects before and after calling the callback.

By creating the ExternalCallEvent we can also offload responsibility from FunctionEvent.