Introduce correct types for Boehm GC
ASObject is derived from class gc, soall objects are allocated in garbage
collected memory.
std containers that stores pointer or smart references to objects uses
traceable_allocator: the GC knows about the memory but will not manage
it. This is good for testing, since only real leaks will be reported by
the GC, but it's actually wrong for production since objects linked in
cycles will not be freed by the GC. All containers should be moved to
gc_allocator.
There is a custom gc_traceable class declared is gc_support.h that is
used for classes that contains object pointers but do not need to be
GCed themselves. For example CairoTokenRenderer.