[URLRequest] Unqualified URLs are relative to "base" attribute or HTML page
Plugin:

If OBJECT or EMBED tag has a "base" attribute, unqualified URLs are
resolved relative to it (if the attribute itself is unqualified, it is
relative to the address of the HTML page). Otherwise URLs are relative
to the address of the containing HTML page.

Standalone Lightspark (similar to AIR):

URLs are relative to the address of the SWF.