Delay binding of symbol class tags until frame initialization
Adds support for symbol class tags that are derived from classes
defined in an external SWF. For example http://edition.cnn.com/video/
needs this.

The root script does stop() and starts loading an external SWF. When
the external file has been loaded and parsed, nextFrame() is called.
The symbol class can not be bound before the initialization of the
second frame because only at that point the base class definition has
been loaded from the external file.