[NetStream] Disable exceptions on FFMpegStreamDecoder stream
Let the decoder handle EOF and other conditions by returning 0 in the
read function.

Exceptions cause problems according to bug 818609.

Another problem I was seeing was that an EOF exception was thrown
while the decoder was looking for metadata on a short video, and the
expected AS events were never dispatched.