remove +500ms in `Sound.position()` (needed to match Adobe’s behaviour)
`Sound.position()` is inaccurate when current sound position is compared to a given time.

To the best of my knowledge, there is a delay in the actual position over 500ms (to one second).

As far as I know, this is an issue (or even a bug) with the official player. 

The closest related info seems to be (at least to me) _Frame subdivision for streaming sound_ from the _SWF File Format Specification_ (https://www.adobe.com/content/dam/acom/en/devnet/pdf/swf-file-format-spec.pdf#page=186 [working link: https://archive.org/download/swf-file-format-spec/swf-file-format-spec.pdf#page=186]).

I requested the addition implemented in f6855b3e1ea94127630328e0caba55ebb1c58496.

Since I may be the only user affected this mismatch and there are ways to avoid the delay, it may be wise to remove the extra time.

Sorry for the noise (it took me years to understand the proper issue).

Many thanks for your help.