[BevelFilter] Optimize bevel filter
We were originally creating 2 seperate bitmaps for both the shadow, and
the highlight. Which meant doing 3 passes of the main bitmap during the
process of applying the bevel filter.

This solves that by computing both the shadow, and highlight pixels within
the main loop, thus, reducing the number of passes of the main bitmap to a
single pass.