[Shader/Filters] Save previous filter output, and use it as the destination
Before, we were only using `g_tex4` as the destination, which never
changes, since it always holds the original DisplayObject texture.

This was causing issues when multiple filters are chained together,
especially if the filter does a blur as one of it's steps.

This change solves that by saving the output of the previous filter, and
making that available for the next filter via `g_tex5`.

Note that for the first filter, we do use `g_tex4`, since `g_tex5` is
already bound to the filter output texture, but for subsequent filters,
we use `g_tex5` from then on.