[Render,DisplayObject] Properly handle masking
Before, we were using the fragment shader to handle masking, and were
never preserving the previous state.

This was causing tons of issues relating to masking order.

This change solves that by keeping track of a mask stack, and then
pushing, or popping when needed.

It also now uses the stencil buffer instead, which is designed for doing
stuff like this.