[TextField] Disable closing tag validation when parsing HTML
HTML allows for mismatched tags, which are treated as that node
extending to the end of the document.

This change makes use of a change to our version of PugiXML that allows
for disabling closing tag validation.

This allows the message textboxes in Super Mario 63 to actually display
their text, albeit with errors.