[DisplayObjectContainer] Add child to list after setting it's variable in `insertLegacyChildAt()`
When calling `_addChildAt()`, it eventually calls the constructor of
that child, but the parent's variable of that child is created after the
constructor is called, when it should be created before calling it.

This more closely matches Flash Player, and allows Adventure Quest
Worlds' client to get past the login screen.