[Rendering] Properly handle transforms
Before this, we were only using the target matrix, with an offset to do
everything, which didn't work in some cases.

This fixes that by using having a transform stack.
And when we need to render a filtered clip, we properly offset that with
the render bounds, and create a new transform stack.

This now works in pretty much every case, and will definitely make
rendering faster, once we get rid of most of our invalidation code.