[MovieClip] Properly handle child clip construction
Before this, we were constructing in `declareFrame()`, and while this
worked, it didn't seem to be the right place to do it.

Instead of that, we construct the children in DisplayObjectContainer's
ActionScript constructor, which makes things both cleaner, and clearer.

It should also be more closer to Flash Player too.