rework rendering DisplayObject to bitmap (again...)
We now render the DisplayObject to a texture on the gpu
and read back the pixels into the BitmapContainer.
This ensures the DisplayObject is rendered in exactly the same way
as in "normal" rendering.
It also makes some stuff like the CairoRenderContext and the
SoftwareInvalidationQueue obsolete (Those will be removed in a separate
commit).

fixes #1096
fixes #1107