[Rendering] Properly handle mask rendering with NanoVG
Before this, we were resorting to rendering masks with cairo, due to
NanoVG using the stencil buffer.

Now, with NanoVG's support for clip paths, we can do it entirely on the
GPU.

There are some new rendering bugs, mainly with Ruffle's
`simple_shapes/masks` test, but that's a different, but related issue.