[timer] Replace previous timer design with an event loop based version
Rather than trying to hook into the existing `TimerThread` code, we now
replace it with one that does the all the timing, while waiting for
platform events.

This not only allows us to get rid of `TimerThread` in most cases, it
makes adding support for other backends/platforms even easier.