[timer] Make timing code even more portable
Before this, we would have different timing implementations for each
platform/backend, which we should be trying to avoid.

Now, we have a single implementation for both updating, and handling
timers.

Currently, only `SDLEventLoop` makes use of it, but later on, we can
port `TimerThread` to make use of this.