[{Args,Option}Parser] Add support for medium options
A medium option is a multi-char short option, and acts similarly to a
long option.

Note that medium options are checked before short options, in order to
prevent a short option with the same starting character from taking
precedent.

This can be useful if want an option with a more descriptive short form.

It can also be useful if you want an option that starts with the same
character as an existing short option.