[test-runner] Add `libtest++`
This is a C++ port of the Rust crate `libtest-mimic`.

The reason this was created was due to the fact that there aren't really
any proper dynamic/runtime testing libraries for C++, most of the
existing testing libraries are for unit tests.

Currently, this version uses a few things that are specific to
Lightspark, such as `tiny_string`, and `Optional`.

Eventually, this should be made more portable, and be moved into it's
own repo.