[AVM1] Add separate version of `Date`
This is because AVM1's version of `Date` has `{g,s}et{,UTC}Year()`,
which are the same as `{g,s}et{,UTC}FullYear()`, but not Y2K compliant.

This fixes a crash with Ruffle's `avm1/date` test.

Note, that the test currently fails due to the date being 'faked' in
their test runner.