[logger] Refactor logging code a bit
This only refactors the part relating to tracing, so it's now using an
interface, that way, it can be overriden by other implementations.

Later on, we can figure out how to do the same for the rest of the
logging code, but for now, it's fine, and finally allows the test runner
to run with multiple test threads.