[test-runner/input] Add/Implement a parser for the LSM file format
This is the initial version of the LSM (Lightspark Movie) file format.

The main idea for this is to eventually allow for Lightspark to have
full TASing (Tool-Assisted Speedrun/Superplay) support.

Like the token parser, this should eventually get put in the main
codebase, more specifically, once work on adding TASing support to
Lightspark has started.

There's absolutely a ton of room for improvement on this, but that can
be done later, for now, it should be enough to make use of it in the
test runner, and get the ball rolling on making our own tests.