[AVM1] Add an AVM1 version of `isEqualStrict()`, and use it in `ActionStrictEquals`
There was a regression with some of the AVM1 XML tests, but I added a
hack fix for it. Eventually, we should refactor the `flash.xml` code to
prevent needing to do that, but for now, it's fine.

This matches Flash Player's behaviour more closely, and allows the
following Ruffle tests to pass.

- avm1/strictequals_swf6
- avm1/strictly_equals