[AVM1] Properly implement case insensitive variable access
In SWF 6, and earlier, property keys are case insensitive, so we have to
support case insensitivity. But the way it was done was completely the
wrong way of going about it.

This change implements support for this more properly, by making use of
`SystemState`'s newly added support for case insensitive string ID
lookup.

This matches Flash Player's behaviour, and allows Ruffle's
`avm1/define_function_case_sensitive` test to pass.

It also fixes the following regressions.

- avm1/propagated_clip_event_order
- timeline/timeline_as2_1

It also causes a regression with Shumway's `avm1/haxe/flocons2` test.