[windows/FileSystem] Don't copy the null terminator in `currentPath()`
oof, that was an issue caused by the fact that I wanted to use
`pathSize` when constructing the `std::wstring`.

The reason is because `GetCurrentDirectoryW()` returns the required
buffer size, **including the null terminator**, if the user supplied
buffer isn't large enough.

This allows the following `FileSystem` tests to pass, in the Windows
version.

- absolute
- canonical