[windows/FileSystem] Set the hard link count to -1 in `StatusFromImpl::fromINFO()`
oof, this one was because of `ghc::filesystem` itself.

`ghc::filesystem` doesn't cache hard links in `directory_entry` on
Windows, it does this with an `#ifdef`.

Rather than add an `#ifdef`, just always set the hard link count to -1,
in the Windows version.

Eventually, we should replace the use of `GetFileAttributesExW()` in
`{symlinkS,s}tatus()` with `GetFileInformationByHandle()`, which does
contain the hard link count.

This allows `FileSystem`'s `DirEntry` test to pass, in the Windows
version.