[windows/Path] Strip root prefix when doing case-sensitive comparisons in `compareImpl()`
This allows the following `Path` tests to pass, in the Windows version.

- compare
- gen

It also finally allows all `FileSystem`, and `Path` tests to pass, in
the Windows version!!!

At least, when symlinks are disabled.

I haven't tested that yet, since the symlink support check in the test
runner assumes you're on Windows 10, or 11, but I've been testing on
Windows 7, which does support symlinks, but requires admin perms to
create.