[Span] Make the endian reading functions more efficient/faster
The previous version, when looked at in Compiler Explorer didn't
generate efficient code in GCC, i.e. it didn't generate a `bswap`, or
`movbe` instruction, but Clang was able to.

This change makes it so the entire thing is done at compile time, which
allows both compilers to generate a `bswap`, or `movbe` instruction.

It should also allow both compilers to generate similarly efficient code
on other instruction sets.